+++ tests/unit/generator/http/test_input.py
@@ -11,7 +11,7 @@ import pytest
 import yaml
 
 from logprep.generator.http.input import EventClassConfig, Input
-from tests.unit.event_generator.http.util import create_test_event_files
+from tests.unit.generator.http.util import create_test_event_files
 
 
 class TestInput: