+++ tests/unit/generator/kafka/test_document_loader.py
@@ -13,7 +13,7 @@ from unittest.mock import MagicMock, patch
 from logprep.generator.kafka.document_loader import DocumentLoader
 from logprep.generator.kafka.kafka_connector import KafkaConsumer
 from logprep.generator.kafka.logger import create_logger
-from tests.testdata.event_generator.kafka.kafka_config_dict import get_config
+from tests.testdata.generator.kafka.kafka_config_dict import get_config
 
 
 class MockedRecord:
@@ -48,9 +48,7 @@ class TestDocumentLoader:
         logger = create_logger(config.logging_level)
         mocked_kafka = MagicMock()
         mocked_kafka.Consumer = MockedConsumer
-        with patch(
-            "logprep.event_generator.kafka.kafka_connector.Consumer", return_value=mocked_kafka
-        ):
+        with patch("logprep.generator.kafka.kafka_connector.Consumer", return_value=mocked_kafka):
             self._document_loader = DocumentLoader(config, logger)
 
     def test_init(self):
@@ -60,14 +58,14 @@ class TestDocumentLoader:
 
     def test_get_from_file(self):
         self._document_loader._source_file = Path(
-            "tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"
+            "tests/testdata/generator/kafka/wineventlog_raw.jsonl"
         )
         documents = self._document_loader._get_from_file()
         assert len(documents) == 500
 
     def test_get_raw_documents_from_file_if_source_file_set(self):
         self._document_loader._source_file = Path(
-            "tests/testdata/event_generator/kafka/wineventlog_raw.jsonl"
+            "tests/testdata/generator/kafka/wineventlog_raw.jsonl"
         )
         documents = self._document_loader._get_raw_documents()
         assert len(documents) == 500